pyproject.toml
[tool.poetry]
name = "family-tree"
version = "0.1.0"
description = ""
authors = ["Alexander Sidorov <sidai@bk.ru>"]

[tool.poetry.dependencies]
python = "3.8.10"
psycopg2-binary = "^2.9.3"
dj-database-url = "^0.5.0"
Django = "3.*"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
